+++ toolium/test/test_visual_test.py
@@ -142,8 +142,10 @@ def test_compare_files_diff_fail(driver_wrapper):
     driver_wrapper.config.set('VisualTests', 'fail', 'true')
     visual = VisualTest(driver_wrapper)
 
-    with pytest.raises(AssertionError):
+    with pytest.raises(AssertionError) as exc:
         visual.compare_files('report_name', file_v1, file_v2, 0)
+    assert str(exc.value) == "The new screenshot '%s' did not match the baseline '%s' " \
+                             "(by a distance of 522.65)" % (file_v1, file_v2)
 
 
 def test_compare_files_size(driver_wrapper):
@@ -158,8 +160,9 @@ def test_compare_files_size_fail(driver_wrapper):
     driver_wrapper.config.set('VisualTests', 'fail', 'true')
     visual = VisualTest(driver_wrapper)
 
-    with pytest.raises(AssertionError):
+    with pytest.raises(AssertionError) as exc:
         visual.compare_files('report_name', file_v1, file_small, 0)
+    assert str(exc.value).startswith('assert (1680, 388) == (1246, 388)')
 
 
 def test_get_img_element(driver_wrapper):
@@ -426,6 +429,10 @@ def test_assert_screenshot_no_enabled_force(driver_wrapper):
     driver_wrapper.config.set('VisualTests', 'enabled', 'false')
     visual = VisualTest(driver_wrapper, force=True)
 
+    # Add v1 baseline image
+    baseline_file = os.path.join(root_path, 'output', 'visualtests', 'baseline', 'firefox', 'screenshot_full.png')
+    shutil.copyfile(file_v1, baseline_file)
+
     # Assert screenshot
     visual.assert_screenshot(None, filename='screenshot_full', file_suffix='screenshot_suffix')
     driver_wrapper.driver.get_screenshot_as_png.assert_called_once_with()
@@ -447,9 +454,10 @@ def test_assert_screenshot_no_enabled_force_fail(driver_wrapper):
     shutil.copyfile(file_v2, baseline_file)
 
     # Assert screenshot
-    with pytest.raises(AssertionError):
+    with pytest.raises(AssertionError) as exc:
         visual.assert_screenshot(None, filename='screenshot_full', file_suffix='screenshot_suffix')
     driver_wrapper.driver.get_screenshot_as_png.assert_called_once_with()
+    assert str(exc.value).endswith("did not match the baseline '%s' (by a distance of 522.65)" % baseline_file)
 
 
 def test_assert_screenshot_full_and_save_baseline(driver_wrapper):
@@ -457,6 +465,7 @@ def test_assert_screenshot_full_and_save_baseline(driver_wrapper):
     with open(file_v1, "rb") as f:
         image_data = f.read()
     driver_wrapper.driver.get_screenshot_as_png.return_value = image_data
+    driver_wrapper.config.set('VisualTests', 'save', 'true')
     visual = VisualTest(driver_wrapper)
 
     # Assert screenshot
@@ -478,6 +487,7 @@ def test_assert_screenshot_element_and_save_baseline(driver_wrapper):
     with open(file_v1, "rb") as f:
         image_data = f.read()
     driver_wrapper.driver.get_screenshot_as_png.return_value = image_data
+    driver_wrapper.config.set('VisualTests', 'save', 'true')
     visual = VisualTest(driver_wrapper)
 
     # Assert screenshot
@@ -531,6 +541,44 @@ def test_assert_screenshot_element_and_compare(driver_wrapper):
     driver_wrapper.driver.get_screenshot_as_png.assert_called_once_with()
 
 
+def test_assert_screenshot_full_without_baseline(driver_wrapper):
+    # Configure driver mock
+    with open(file_v1, "rb") as f:
+        image_data = f.read()
+    driver_wrapper.driver.get_screenshot_as_png.return_value = image_data
+    driver_wrapper.config.set('VisualTests', 'fail', 'true')
+    visual = VisualTest(driver_wrapper)
+
+    # Assert screenshot
+    with pytest.raises(AssertionError) as exc:
+        visual.assert_screenshot(None, filename='screenshot_full', file_suffix='screenshot_suffix')
+    driver_wrapper.driver.get_screenshot_as_png.assert_called_once_with()
+    baseline_file = os.path.join(root_path, 'output', 'visualtests', 'baseline', 'firefox', 'screenshot_full.png')
+    assert str(exc.value) == 'Baseline file not found: %s' % baseline_file
+
+
+def test_assert_screenshot_element_without_baseline(driver_wrapper):
+    # Add baseline image
+    driver_wrapper.driver.execute_script.return_value = 0  # scrollX=0 and scrollY=0
+    driver_wrapper.config.set('VisualTests', 'fail', 'true')
+    visual = VisualTest(driver_wrapper)
+
+    # Create element mock
+    web_element = get_mock_element(x=250, y=40, height=40, width=300)
+
+    # Configure driver mock
+    with open(file_v1, "rb") as f:
+        image_data = f.read()
+    driver_wrapper.driver.get_screenshot_as_png.return_value = image_data
+
+    # Assert screenshot
+    with pytest.raises(AssertionError) as exc:
+        visual.assert_screenshot(web_element, filename='screenshot_elem', file_suffix='screenshot_suffix')
+    driver_wrapper.driver.get_screenshot_as_png.assert_called_once_with()
+    baseline_file = os.path.join(root_path, 'output', 'visualtests', 'baseline', 'firefox', 'screenshot_elem.png')
+    assert str(exc.value) == 'Baseline file not found: %s' % baseline_file
+
+
 def test_assert_screenshot_mobile_resize_and_exclude(driver_wrapper):
     # Create elements mock
     driver_wrapper.driver.execute_script.return_value = 0  # scrollX=0 and scrollY=0
@@ -544,6 +592,7 @@ def test_assert_screenshot_mobile_resize_and_exclude(driver_wrapper):
 
     # Update conf and create a new VisualTest instance
     driver_wrapper.config.set('Driver', 'type', 'ios')
+    driver_wrapper.config.set('VisualTests', 'save', 'true')
     visual = VisualTest(driver_wrapper)
 
     # Assert screenshot
@@ -577,6 +626,7 @@ def test_assert_screenshot_mobile_web_resize_and_exclude(driver_wrapper):
     # Update conf and create a new VisualTest instance
     driver_wrapper.config.set('Driver', 'type', 'ios')
     driver_wrapper.config.set('AppiumCapabilities', 'browserName', 'safari')
+    driver_wrapper.config.set('VisualTests', 'save', 'true')
     visual = VisualTest(driver_wrapper)
 
     # Assert screenshot
@@ -596,11 +646,13 @@ def test_assert_screenshot_mobile_web_resize_and_exclude(driver_wrapper):
 
 def test_assert_screenshot_str_threshold(driver_wrapper):
     visual = VisualTest(driver_wrapper)
-    with pytest.raises(TypeError):
+    with pytest.raises(TypeError) as exc:
         visual.assert_screenshot(None, 'screenshot_full', threshold='name')
+    assert str(exc.value) == 'Threshold must be a number between 0 and 1: name'
 
 
 def test_assert_screenshot_greater_threshold(driver_wrapper):
     visual = VisualTest(driver_wrapper)
-    with pytest.raises(TypeError):
+    with pytest.raises(TypeError) as exc:
         visual.assert_screenshot(None, 'screenshot_full', threshold=2)
+    assert str(exc.value) == 'Threshold must be a number between 0 and 1: 2'