+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ v1.6.2.dev0
 
 - Fix to allow step's text (context.text) declaration into dynamic environment sections
 - Add `ssl` config property in [Server] section to allow using https in Selenium Grid url
+- Visual testing comparison must fail when baseline does not exist and save mode is disabled
 
 v1.6.1
 ------